Deleted ElasticSearchShortcuts and moved methods to ElasticSearchAdminService to be exposed to applications for convenience
